# SslHandler aggregation of plaintext data on write