… handler/src/test/java/io/netty/handler/ssl/SslHandlerTest.java
… transport/src/main/java/io/netty/channel/CoalescingBufferQueue.java
… transport/src/main/java/io/netty/channel/DelegatingChannelPromiseNotifier.java