@Scottmitch 2017-05-09
- This PR depends upon PR https://github.com/netty/netty/pull/6721

See some some benchmark results below. The numbers look very similar before/after. The microbenchmark may not suffer much from lower goodput because the handlers are directly wired together. I would like to test in a more realistic scenario where network latency/throughput may be limiting.

```
Before
Benchmark                           (bufferType)                               (cipher)  (messageSize)  (numWrites)  (sslProvider)   Mode  Cnt       Score       Error  Units
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64            1        OPENSSL  thrpt   20  442428.122 ± 35762.490  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64           10        OPENSSL  thrpt   20   52367.433 ±  2414.133  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096            1        OPENSSL  thrpt   20  179066.139 ±  6834.309  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096           10        OPENSSL  thrpt   20   19753.676 ±   501.900  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64            1        OPENSSL  thrpt   20  533367.506 ±  9579.907  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64           10        OPENSSL  thrpt   20   53857.118 ±  1419.701  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096            1        OPENSSL  thrpt   20  204300.938 ±  3423.820  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096           10        OPENSSL  thrpt   20   20458.150 ±   929.823  ops/s

After
Benchmark                           (bufferType)                               (cipher)  (messageSize)  (numWrites)  (sslProvider)   Mode  Cnt       Score       Error  Units
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64            1        OPENSSL  thrpt   20  479451.492 ± 13056.271  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64           10        OPENSSL  thrpt   20   49086.053 ±  2919.347  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096            1        OPENSSL  thrpt   20  183940.496 ±  4922.291  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096           10        OPENSSL  thrpt   20   19421.915 ±   490.627  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64            1        OPENSSL  thrpt   20  515227.532 ± 13316.443  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64           10        OPENSSL  thrpt   20   54458.104 ±  1164.469  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096            1        OPENSSL  thrpt   20  199174.137 ±  6294.416  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096           10        OPENSSL  thrpt   20   20128.592 ±   737.846  ops/s
```