@Scottmitch 2017-05-10
See below for preliminary benchmark results. The benefits are more visible when many smaller packets are written, and it looks like performance is on par when the input buffers are larger. Considerations:
- We may do more allocations/copies during aggregation ... will the behaviors change in applications where memory/allocator bandwidth behavior is different.
- Even though the performance looks the same with larger buffers, this test is not a good reflection of goodput because there is no network involved. It would be interesting to do analysis when the network throughput is limited or loss maybe present.

```
With aggregation
Benchmark                           (bufferType)                               (cipher)  (messageSize)  (numWrites)  (sslProvider)   Mode  Cnt       Score       Error  Units
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64           10        OPENSSL  thrpt   10  216645.213 ± 18545.528  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64          100        OPENSSL  thrpt   10   41643.772 ±  2106.882  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           1024           10        OPENSSL  thrpt   10   79265.317 ±  6932.062  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           1024          100        OPENSSL  thrpt   10    8754.069 ±   471.745  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096           10        OPENSSL  thrpt   10   23823.842 ±  1346.592  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096          100        OPENSSL  thrpt   10    2340.937 ±   107.498  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           8192           10        OPENSSL  thrpt   10   12306.394 ±   655.865  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           8192          100        OPENSSL  thrpt   10    1181.429 ±   116.225  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64           10        OPENSSL  thrpt   10  167119.477 ±  9279.883  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64          100        OPENSSL  thrpt   10   37818.274 ±  2197.255  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           1024           10        OPENSSL  thrpt   10   79088.807 ±  2515.297  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           1024          100        OPENSSL  thrpt   10    7619.249 ±   337.530  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096           10        OPENSSL  thrpt   10   22146.381 ±  1184.731  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096          100        OPENSSL  thrpt   10    2149.471 ±   184.393  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           8192           10        OPENSSL  thrpt   10   11958.406 ±   994.900  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           8192          100        OPENSSL  thrpt   10    1137.731 ±    23.498  ops/s

Without aggregation
Benchmark                           (bufferType)                               (cipher)  (messageSize)  (numWrites)  (sslProvider)   Mode  Cnt      Score      Error  Units
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64           10        OPENSSL  thrpt   10  50482.154 ± 4090.253  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64          100        OPENSSL  thrpt   10   2617.450 ±  133.296  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           1024           10        OPENSSL  thrpt   10  36896.912 ± 2481.016  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           1024          100        OPENSSL  thrpt   10   2925.621 ±  476.696  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096           10        OPENSSL  thrpt   10  19105.415 ± 1659.986  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096          100        OPENSSL  thrpt   10   1719.703 ±   70.245  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           8192           10        OPENSSL  thrpt   10  12056.736 ±  384.316  ops/s
SslHandlerEchoBenchmark.wrapUnwrap          HEAP  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           8192          100        OPENSSL  thrpt   10   1070.577 ±   43.189  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64           10        OPENSSL  thrpt   10  53221.344 ± 2208.686  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256             64          100        OPENSSL  thrpt   10   2648.294 ±  165.110  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           1024           10        OPENSSL  thrpt   10  39705.972 ± 1448.021  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           1024          100        OPENSSL  thrpt   10   3273.145 ±  158.115  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096           10        OPENSSL  thrpt   10  20943.006 ± 1338.166  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           4096          100        OPENSSL  thrpt   10   1837.158 ±   89.656  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           8192           10        OPENSSL  thrpt   10  12595.608 ± 1191.738  ops/s
SslHandlerEchoBenchmark.wrapUnwrap        DIRECT  TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256           8192          100        OPENSSL  thrpt   10   1124.542 ±   63.260  ops/s
```